# com.sap.cloud.lm.sl.cf.api/test_rest-mtarest-1.0.0-swagger.yaml
...
swagger: "2.0"
info:
  description: "This is an API for the deploy-service(MTA) application"
  version: "1.0.0"
  title: "MTA REST API"
  contact:
    email: "DL_563CD7577BCF844DBD000037@exchange.sap.corp"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
# host: "deploy-service.bosh-lite.com"
# basePath: "/api/v1/spaces/{space_guid}"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- oauth2: []
paths:
...
  /files:
    get:
      description: "Retrieves all Multi-Target Application files\n"
      operationId: "GetMtaFiles"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/FileMetadata"
    post:
      description: "Uploads an Multi Target Application file\n"
      operationId: "UploadMtaFile"
      consumes: 
        - "multipart/form-data"
      parameters:
      - name: "file"
        in: "formData"
        required: true
        type: "file"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/FileMetadata"
...
securityDefinitions:
  oauth2:
    type: "oauth2"
    tokenUrl: "https://uaa.bosh-lite.com/oauth/token"
    flow: "password"
    scopes: {}
definitions:
  FileMetadata:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      size:
        type: "number"
      digest:
        type: "string"
      digestAlgorithm:
        type: "string"
      space:
        type: "string"
...